adding binary_type c++ and go for base.py

+++ release_tester/arangodb/installers/base.py
@@ -10,6 +10,7 @@ from pathlib import Path
 from abc import abstractmethod, ABC
 import yaml
 import shutil
+import time
 from arangodb.instance import ArangodInstance
 from tools.asciiprint import ascii_print, print_progress as progress
 
@@ -33,7 +34,7 @@ if is_windows:
 ## helper classes
 class BinaryDescription():
     """ describe the availability of an arangodb binary and its properties """
-    def __init__(self, path, name, enter, strip, vmin, vmax, sym):
+    def __init__(self, path, name, enter, strip, vmin, vmax, sym, binary_type):
         global winver
         self.path = path / (name + extension)
         self.enterprise = enter
@@ -41,6 +42,7 @@ class BinaryDescription():
         self.version_min = vmin
         self.version_max = vmax
         self.symlink = sym
+        self.binary_type = binary_type
 
         for attribute in (
                 self.path,
@@ -48,7 +50,8 @@ class BinaryDescription():
                 self.stripped,
                 self.version_min,
                 self.version_max,
-                self.symlink
+                self.symlink,
+                self.binary_type
         ):
             if attribute is None:
                 logging.error("one of the given args is null")
@@ -63,6 +66,7 @@ class BinaryDescription():
         version_min: {0.version_min}
         version_max: {0.version_max}
         symlink:     {0.symlink}
+        binary_type: {0.binary_type}
         """.format(self)
 
 
@@ -96,58 +100,72 @@ class BinaryDescription():
 
     def check_stripped_mac(self):
         """ Checking stripped status of the arangod """
-        print('stripped command invoked on: ' + str(self.path))
-        # finding out the file size before stripped cmd invoked
-        beforStripped = self.path.stat().st_size
-        # print('Before stripped File size in bytes: ', beforStripped)
+        if self.binary_type == 'c++':  
+            # print('stripped command invoked on: ' + str(self.path))
+            # finding out the file size before stripped cmd invoked
+            beforStripped = self.path.stat().st_size
+            # print('before ' + str(beforStripped))
 
-        to_file = Path('/tmp/test_whether_stripped')
-        shutil.copy(str(self.path), str(to_file))
-        
-        # invoke the strip command on file_path
-        cmd = ['strip', str(to_file)]
-        # print(cmd)
-        proc = subprocess.Popen(cmd, bufsize=-1,
-                            stderr=subprocess.PIPE,
-                            stdin=subprocess.PIPE)
+            to_file = Path('/tmp/test_whether_stripped')
+            shutil.copy(str(self.path), str(to_file))
+            
+            # invoke the strip command on file_path
+            cmd = ['strip', str(to_file)]
+            # print(cmd)
+            proc = subprocess.Popen(cmd, bufsize=-1,
+                                stderr=subprocess.PIPE,
+                                stdin=subprocess.PIPE)
 
-        # check the size of copied file after stripped
-        afterStripped = to_file.stat().st_size
-        # print('After stripped file size in bytes: ', afterStripped)
-        
-        # checking both output size 
-        if beforStripped == afterStripped:
-            print('Stripped status: binary is stripped')
+            # check the size of copied file after stripped
+            afterStripped = to_file.stat().st_size
+            # print('after ' + str(afterStripped))
+            
+            # checking both output size 
+            if beforStripped == afterStripped:
+                # print('Stripped status: binary is stripped')
+                return True
+            else:
+                # print('Stripped status: binary is not stripped')
+                return False
+            
+            
+            if to_file.is_file():
+                # invoke the delete command on file_path
+                to_file.unlink(str(to_file))
+      
+            else:
+                print('stripped file not found')
         else:
-            print('Stripped status: binary is not stripped')
-        
-        
-        if to_file.is_file():
-            # invoke the delete command on file_path
-            to_file.unlink(str(to_file))
-            # print(str(self.path) + '\'s copy file deleted after stripped check \n')
-        else:
-            print('stripped file not found')
-        
+            # print('Strip check skipped for GO binary')
+            return True
+            
     
     def check_stripped(self):
         """ check whether this file is stripped (or not) """
         # checking stripped state for macos
+        time.sleep(1)
         output = run_file_command(self.path)
         macver = platform.mac_ver()
         if macver[0]:
-            if self.check_stripped_mac() and not self.stripped:
+            # print('binary_type: ' + str(self.binary_type))
+            is_stripped = self.check_stripped_mac()
+            # print(is_stripped)
+            if not is_stripped and self.stripped:
                  raise Exception("expected " + str(self.path) +
-                            " to be stripped, but its not: " + output)
+                            " to be stripped, but it is not: " + output)
+            
+            # if is_stripped and not self.stripped:
+            #     raise Exception("expected " + str(self.path) +
+            #                 " not to be stripped, but it is stripped: " + output)
 
         else:
             if self.stripped and output.find(', stripped') < 0:
                 raise Exception("expected " + str(self.path) +
-                                " to be stripped, but its not: " + output)
+                                " to be stripped, but it is not: " + output)
 
             if not self.stripped and output.find(', not stripped') < 0:
                 raise Exception("expected " + str(self.path) +
-                                " to be stripped, but its not: " + output)
+                                " not to be stripped, but it is stripped: " + output)
 
 
     def check_symlink(self):
@@ -296,71 +314,72 @@ class InstallerBase(ABC):
             False, True, "1.0.0", "4.0.0", [
                 self.cfg.real_sbin_dir / 'arango-init-database',
                 self.cfg.real_sbin_dir / 'arango-secure-installation'
-            ]))
+            ], 'c++'))
 
         # symlink only for MMFILES
         self.arango_binaries.append(BinaryDescription(
             self.cfg.real_sbin_dir, 'arangod',
             False, True, "1.0.0", "3.6.0", [
                 self.cfg.real_bin_dir / ('arango-dfdb' + extension)
-            ]))
+            ], 'c++'))
 
         self.arango_binaries.append(BinaryDescription(
             self.cfg.real_bin_dir, 'arangosh',
             False, True, "1.0.0", "4.0.0", [
                 self.cfg.real_bin_dir / ('arangoinspect' + extension)
-            ]))
+            ], 'c++'))
 
         self.arango_binaries.append(BinaryDescription(
             self.cfg.real_bin_dir, 'arangoexport',
-            False, True, "1.0.0", "4.0.0", []))
+            False, True, "1.0.0", "4.0.0", [], 'c++'))
 
         self.arango_binaries.append(BinaryDescription(
             self.cfg.real_bin_dir, 'arangoimport',
             False, True, "1.0.0", "4.0.0", [
                 self.cfg.real_bin_dir / ('arangoimp' + extension)
-            ]))
+            ], 'c++'))
 
         self.arango_binaries.append(BinaryDescription(
             self.cfg.real_bin_dir, 'arangodump',
-            False, True, "1.0.0", "4.0.0", []))
+            False, True, "1.0.0", "4.0.0", [], 'c++'))
 
         self.arango_binaries.append(BinaryDescription(
             self.cfg.real_bin_dir, 'arangorestore',
-            False, True, "1.0.0", "4.0.0", []))
+            False, True, "1.0.0", "4.0.0", [], 'c++'))
 
         self.arango_binaries.append(BinaryDescription(
             self.cfg.real_bin_dir, 'arangobench',
-            False, True, "1.0.0", "4.0.0", []))
+            False, True, "1.0.0", "4.0.0", [], 'c++'))
 
         self.arango_binaries.append(BinaryDescription(
             self.cfg.real_bin_dir, 'arangovpack',
-            False, True, "1.0.0", "4.0.0", []))
+            False, True, "1.0.0", "4.0.0", [], 'c++'))
 
         #starter
         self.arango_binaries.append(BinaryDescription(
             self.cfg.real_bin_dir, 'arangodb',
-            False, False, "1.0.0", "4.0.0", []))
+            False, False, "1.0.0", "4.0.0", [], 'go'))
 
-        ## enterprise
+        # enterprise
         self.arango_binaries.append(BinaryDescription(
             self.cfg.real_bin_dir, 'arangobackup',
-            True, True, "1.0.0", "4.0.0", []))
+            True, True, "1.0.0", "4.0.0", [], 'c++'))
 
         self.arango_binaries.append(BinaryDescription(
             self.cfg.real_sbin_dir, 'arangosync',
             True, False, "1.0.0", "4.0.0", [
                 self.cfg.real_bin_dir / ('arangosync' + extension)
-            ]))
+            ], 'go'))
 
         self.arango_binaries.append(BinaryDescription(
             self.cfg.real_sbin_dir, 'rclone-arangodb',
-            True, True, "1.0.0", "4.0.0", []))
+            True, True, "1.0.0", "4.0.0", [], 'c++'))
 
     def check_installed_files(self):
         """ check for the files whether they're installed """
+        print('Invoking strip checking')
         for binary in self.arango_binaries:
-            # progress("S" if binary.stripped else "s")
+            progress("S" if binary.stripped else "s")
             binary.check_installed(self.cfg.version,
                                    self.cfg.enterprise,
                                    self.check_stripped,
